% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_delim.R
\name{identify_delim}
\alias{identify_delim}
\title{Identifica potenciales delimitadores del archivo}
\usage{
identify_delim(path, delims = c("\\t", ",", " ", "|", ";"), n = 10,
  comment = "#", skip = 0, quote = "\\"")
}
\arguments{
\item{path}{ruta del archivo}

\item{delims}{un vector de delimitadores a tratar}

\item{n}{numero de filas a mirar en el archivo para determinar los delimitadores}

\item{comment}{identica las lineas que son comentarios}

\item{skip}{numero de lineas a evitar al inicio}

\item{quote}{configura el tipo de texto para frases}
}
