% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_datasets.R
\name{tt_datasets}
\alias{tt_datasets}
\title{Bases de datos disponibles}
\usage{
tt_datasets(year)
}
\arguments{
\item{year}{dato numerico que representa cada periodo anual. Dejarlo vacio para el periodo mas actual.}
}
\description{
lista de bases de datos completa cada periodo
}
